fix: success modal error
