+++ tailwind.css
@@ -31,7 +31,7 @@
     --warning-200: 255 231 173;
     --warning-100: 255 243 214;
 
-    --error-400: 255 69 38;
+    --error-400: 222 69 38;
     --error-300: 230 115 92;
     --error-200: 242 181 168;
     --error-100: 248 218 212;
