+++ web-marketplace/src/components/organisms/PostFlow/PostFlow.utils.tsx
@@ -1,21 +1,4 @@
-import { Trans } from '@lingui/macro';
-
-import { TextButton } from 'web-components/src/components/buttons/TextButton';
-import { Item } from 'web-components/src/components/modal/TxModal';
-import { CardItemValue } from 'web-components/src/components/modal/TxModal.CardItemValue';
-import { Subtitle } from 'web-components/src/components/typography';
-import { truncate } from 'web-components/src/utils/truncate';
-
-import { getHashUrl } from 'lib/block-explorer';
-import {
-  Post,
-  PostFile,
-} from 'lib/queries/react-query/registry-server/getPostQuery/getPostQuery.types';
-
-import { Link } from 'components/atoms';
-
 import { PostFormSchemaType } from '../PostForm/PostForm.schema';
-import { FILE_NAMES, HASH, PROJECT } from './PostFlow.constants';
 
 export const getFiles = (files?: PostFormSchemaType['files']) => {
   return files?.map(file => {
@@ -30,123 +13,5 @@ export const getFiles = (files?: PostFormSchemaType['files']) => {
   });
 };
 
-type GetSuccessModalContentParams = {
-  createdPostData?: Post | null;
-  projectSlug?: string | null;
-  projectId?: string;
-  offChainProjectId?: string;
-  projectName?: string;
-  anchorTxHash?: string;
-  attestTxHash?: string;
-  signingFailed?: boolean;
-};
-export const getSuccessModalContent = ({
-  createdPostData,
-  projectSlug,
-  projectId,
-  offChainProjectId,
-  projectName,
-  anchorTxHash,
-  attestTxHash,
-  signingFailed,
-}: GetSuccessModalContentParams) => {
-  const iri = createdPostData?.iri;
-  const files = createdPostData?.contents?.files as PostFile[] | undefined;
-  const filesUrls = createdPostData?.filesUrls;
-
-  const projectUrl = `/project/${
-    projectSlug ?? projectId ?? offChainProjectId
-  }`;
-  const cardItems: Item[] = [
-    {
-      label: PROJECT,
-      value: {
-        name: projectName ?? projectId ?? offChainProjectId,
-        url: projectUrl,
-      },
-    },
-  ];
-  if (files && files.length && filesUrls && filesUrls.length) {
-    cardItems.unshift({
-      label: FILE_NAMES,
-      value:
-        files.map(file => ({
-          name: file.name as string,
-          url: filesUrls[file.iri],
-        })) || [],
-    });
-  }
-
-  const hashValue = [];
-  if (anchorTxHash) {
-    hashValue.push({
-      component: (
-        <div className="flex">
-          <TextButton
-            textSize="sm"
-            className="font-extrabold text-grey-400 hover:text-grey-400 pr-3"
-          >
-            <Trans>anchor</Trans>:
-          </TextButton>
-          <CardItemValue
-            size="sm"
-            mobileSize="sm"
-            value={{
-              name: truncate(anchorTxHash),
-              url: getHashUrl(anchorTxHash),
-            }}
-            linkComponent={Link}
-          />
-        </div>
-      ),
-    });
-  } else {
-    hashValue.push({
-      component: (
-        <Subtitle size="lg" className="font-normal text-grey-400 italic">
-          <Trans>Anchoring in progress</Trans>
-        </Subtitle>
-      ),
-    });
-  }
-
-  if (attestTxHash) {
-    hashValue.push({
-      component: (
-        <div className="flex">
-          <TextButton
-            textSize="sm"
-            className="font-extrabold text-grey-400 hover:text-grey-400 pr-3"
-          >
-            <Trans>attest</Trans>:{' '}
-          </TextButton>
-          <CardItemValue
-            size="sm"
-            mobileSize="sm"
-            value={{
-              name: truncate(attestTxHash),
-              url: getHashUrl(attestTxHash),
-            }}
-            linkComponent={Link}
-          />
-        </div>
-      ),
-    });
-  } else if (signingFailed) {
-    hashValue.push({
-      component: (
-        <Subtitle size="lg" className="text-error-400">
-          <Trans>Signing failed</Trans>
-        </Subtitle>
-      ),
-    });
-  }
-
-  if (hashValue.length > 0)
-    cardItems.push({
-      label: HASH,
-      value: hashValue,
-    });
-
-  return { cardItems, buttonLink: `/post/${iri}` };
-};
+export const timer = (ms: number): Promise<NodeJS.Timeout> =>
+  new Promise(res => setTimeout(res, ms));
